+++ lib/kettle-soup-cover.rb
@@ -1,4 +1,4 @@
-# USAGE:
+# rubocop:disable Naming/FileName# USAGE:
 # In your `spec/spec_helper.rb`,
 # just prior to loading the library under test:
 #
@@ -10,13 +10,8 @@
 #   SimpleCov.start
 #
 
-# External gems
-require "version_gem"
-
-require_relative "kettle/soup/cover/version"
-
+# For technical reasons, if we move to Zeitwerk, this cannot be require_relative.
+#   See: https://github.com/fxn/zeitwerk#for_gem_extension
+# Hook for other libraries to load this library (e.g. via bundler)
 require "kettle/soup/cover"
-
-Kettle::Soup::Cover::Version.class_eval do
-  extend VersionGem::Basic
-end
+# rubocop:enable Naming/FileName