+++ lib/kettle/soup/cover.rb
@@ -14,7 +14,14 @@
 #   SimpleCov.start
 #
 
+# Standard Lib
 require "rbconfig"
+
+# External gems
+require "version_gem"
+
+# This gem
+require_relative "cover/version"
 
 module Kettle
   module Soup
@@ -173,4 +180,8 @@ module Kettle
       end
     end
   end
+end
+
+Kettle::Soup::Cover::Version.class_eval do
+  extend VersionGem::Basic
 end